### 2022-06-18 #4485 @arfon
@perdelt – many thanks for your submission to JOSS. I'm flagging this for a scope query by our editorial board as it't not clear to me on first inspection that this software meets our definition of [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software). This process will likely take a week or so.

In the meantime, please take a look at the errors for your paper, namely that one of the authors appears to be missing an affiliation: https://github.com/openjournals/joss-papers/runs/6950440027?check_suite_focus=true#step:3:86

You can find an example paper here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
