### 2022-06-22 #4485 @perdelt
@arfon - many thanks for your comment. I could improve the statement of need to stress that performance benchmarking is an important research domain of DBMS
