### 2022-06-27 #4485 @perdelt
@gkthiruvathukal many thanks for contacting me.
There are a lot of people on the list understanding Python and who are interested in DBMS, testing and reproducible research. How many names would be appropriate?
To pick a few, I think suitable would be: simon-lewis,erik-whiting,ChristopherHaydenTodd,remram44
