### 2022-06-28 9746e31
Inspector: workload dict without metrics and reporting
