### 2022-06-28 5658a1c
Inspector: workload dict without metrics and reporting
