### 2022-06-28 58d9be0
Inspector: workload dict without metrics and reporting
