+++ dbmsbenchmarker/inspector.py
@@ -145,6 +145,11 @@ class inspector():
         self.benchmarks.computeTimerRun()
         self.benchmarks.computeTimerSession()
         self.e = evaluator.evaluator(self.benchmarks, load=load, force=True)
+        self.workload = copy.deepcopy(self.e.evaluation['general'])
+        # remove metrics
+        del(self.workload['loadingmetrics'])
+        del(self.workload['streamingmetrics'])
+        del(self.workload['reporting'])
     def get_experiment_list_queries(self):
         # list of successful queries
         return self.benchmarks.listQueries()
@@ -261,12 +266,7 @@ class inspector():
             return self.e.evaluation['query']
     def get_experiment_workload_properties(self):
         # dict of workload properties
-        workload = copy.deepcopy(self.e.evaluation['general'])
-        # remove metrics
-        del(workload['loadingmetrics'])
-        del(workload['streamingmetrics'])
-        del(workload['reporting'])
-        return workload
+        return self.workload
     #def get_measures(self, numQuery, timer, warmup=0, cooldown=0):
     def get_timer(self, numQuery, timer, warmup=0, cooldown=0):
         # dataframe of dbms x measures
