### 2022-06-29 e9c0f3c
V0.12.1 (#93)
* Prepare next release

* Inspector: workload dict without metrics and reporting

* Docs: JOSS intro
