### 2022-06-29 #4485 @simon-lewis
> @simon-lewis are you willing to contribute a review for this JOSS submission?

Happy to review. Let me know if to proceed. Thanks.
