### 2022-06-30 7459ed5
Inspector: workload dict without metrics and reporting - catch missing values
