### 2022-07-25 8514649
Dashboard: Client from Microservices
