### 2022-07-30 #4628 @editorialbot
[REVIEW]: DBMS-Benchmarker: Benchmark and Evaluate DBMS in Python
**Submitting author:** @perdelt (<a href="http://orcid.org/0000-0002-3359-2386">Patrick Erdelt</a>)
**Repository:** https://github.com/Beuth-Erdelt/DBMS-Benchmarker
**Branch with paper.md** (empty if default branch):
**Version:** v0.11.22
