### 2022-07-31 #4628 @erik-whiting
@gkthiruvathukal has the citation format for the journal changed in the last few months? In reading the paper, I noticed the use of "c.f." throughout and I am unsure what it means. It appears before some citations or after others, or sometimes not at all. I'm pretty confused by it, can you confirm if it is something specific to JOSS? Sorry if this is a silly quesiton.

@perdelt if you know the answer to this, please feel free to chime in.
