### 2022-07-31 #4628 @perdelt
@erik-whiting Thank you for your comment!
This is intended to mark references that are not a source of citation, but rather invite the reader to compare information provided elsewhere.
https://tinyurl.com/mwd2ak3c
I see it's actually used a lot here, so I might remove a few.
