### 2022-07-31 #4628 @gkthiruvathukal
@erik-whiting, @perdelt's explanation is correct. In general, I would like to suggest minimizing use of these additional references. These would probably make more sense for an expanded work, perhaps an article that eventually cites this software, if accepted by JOSS, of course.
