+++ paper.md
@@ -42,22 +42,22 @@ All of these products have unique characteristics, special use cases, advantages
 In order to be able to verify and ensure the performance measurement, we want to be able to create and repeat benchmarking scenarios.
 Repetition and thorough evaluation are crucial, in particular in the age of Cloud-based systems with it's diversity of hardware configurations, @Raasveldt2018FBC32099503209955, @DBLPconfsigmodKerstenKZ18, @KounevLK20.
 
-Performance benchmarking of DBMS has a broad audience. It is used *by DBMS developers to evaluate their work and to find out which algorithm works best in which situation. Benchmarks are used by (potential) customers to evaluate what system or hardware to buy or rent. Benchmarks are used by administrators to find bottlenecks and adjust configurations. Benchmarks are used by users to compare semantically equivalent queries and to find the best formulation alternative* @10.1007/978-3-030-84924-5_6.
+Performance benchmarking of DBMS has a broad audience. It is used *by DBMS developers to evaluate their work and to find out which algorithm works best in which situation. Benchmarks are used by (potential) customers to evaluate what system or hardware to buy or rent. Benchmarks are used by administrators to find bottlenecks and adjust configurations. Benchmarks are used by users to compare semantically equivalent queries and to find the best formulation alternative*, @10.1007/978-3-030-84924-5_6.
 
 There thus is widespread need for a tool to support the repetition and reproducibility of benchmarking situations, and that is capable of connecting to all these systems.
 
 There is also need for a tool that will help with the statistical and interactive analysis of the results.
 We want to use Python as the common Data Science language, since
 *it is a mature language programming, easy for the newbies, and can be used as a specific platform for data scientists, thanks to its large ecosystem of scientific libraries and its high and vibrant community* @series/utcs/IgualS17.
-This helps to implement the tool into a pipeline, for example to make use of closed-loop benchmarking situations @10114533389063338912 or to closely inspect parts of queries @DBLPconfsigmodKerstenKZ18.
+This helps to implement the tool into a pipeline, for example to make use of closed-loop benchmarking situations, @10114533389063338912, or to closely inspect parts of queries, @DBLPconfsigmodKerstenKZ18.
 It also allows to use common and sophisticated tools to inspect and evaluate the results.
 To name a few:
-Pandas for statistical evaluation of tabular data @reback2020pandas, @mckinney-proc-scipy-2010,
-Scipy for scientific investigation of data @2020SciPy-NMeth,
-IPython and Jupyter notebooks for interactive analysis and display of results @Kluyver2016jupyter,
-Matplotlib and Seaborn for visual analysis @Hunter:2007, @Waskom2021,
+Pandas for statistical evaluation of tabular data, @reback2020pandas, @mckinney-proc-scipy-2010,
+Scipy for scientific investigation of data, @2020SciPy-NMeth,
+IPython and Jupyter notebooks for interactive analysis and display of results, @Kluyver2016jupyter,
+Matplotlib and Seaborn for visual analysis, @Hunter:2007, @Waskom2021,
 or even Machine Learning tools.
-Moreover Python is currently the most popular computer language @PYPL, @TIOBE.
+Moreover Python is currently the most popular computer language, @PYPL, @TIOBE.
 
 To our knowledge there is no other such tool, c.f. results of @10.1007/978-3-319-67162-8_12, @10.1007/978-3-030-12079-5_4.
 There are other tools like Apache JMeter, HammerDB, Sysbench, OLTPBench, that provide very nice features, but none fitting these needs.
