### 2022-08-01 #4628 @simon-lewis
@gkthiruvathukal @perdelt @erik-whiting  -
Hi team. Just weighing in on this one. In my opinion, "cf." tends to be used to cite _alternative_ sources, even contradictory ones. It means "compare with" so typically used after another citation, to say "but also check this one out". Hope that helps.

I've started my review so will feed back in more detail.

Like @erik-whiting , I will try to avoid commenting too deeply on grammar, unless it really is a question of meaning / understanding.
