+++ paper.bib
@@ -276,7 +276,8 @@ pages = {188--199},
 numpages = {12},
 keywords = {cloud computing, resource contention, performance testing, non- parametric statistics},
 location = {Tallinn, Estonia},
-series = {ESEC/FSE 2019}
+series = {ESEC/FSE 2019},
+doi = {10.1145/3338906.3338912},
 }
 
 @inproceedings{DBLPconfsigmodKerstenKZ18,
@@ -292,6 +293,7 @@ series = {ESEC/FSE 2019}
   publisher = {{ACM}},
   year      = {2018},
   url       = {https://doi.org/10.1145/3209950.3209951},
+  doi       = {10.1145/3209950.3209951},
   timestamp = {Mon, 12 Aug 2019 13:49:51 +0200},
   biburl    = {https://dblp.org/rec/conf/sigmod/KerstenKZ18.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
