+++ paper.bib
@@ -156,7 +156,8 @@ Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and A
 Editor = {F. Loizides and B. Schmidt},
 Organization = {IOS Press},
 Pages = {87 - 90},
-Year = {2016}
+Year = {2016},
+doi = {10.3233/978-1-61499-649-1-87}
 }
 
 @Article{Hunter:2007,
@@ -219,6 +220,7 @@ Year = {2016}
   author = {Igual, Laura and Seguí, Santi},
   biburl = {https://www.bibsonomy.org/bibtex/2df957256a471cf3626c5a280424a2b9a/dblp},
   ee = {https://doi.org/10.1007/978-3-319-50017-1},
+  doi = {10.1007/978-3-319-50017-1},
   interhash = {1f6d348e064db5ee3a118549b35cbb04},
   intrahash = {df957256a471cf3626c5a280424a2b9a},
   isbn = {978-3-319-50017-1},
