### 2022-08-01 0e7f53f
Joss: Statement of need footnotes
