+++ paper.bib
@@ -302,5 +302,25 @@ doi = {10.1145/3338906.3338912},
 }
 
 
+@misc{DBDBIO,
+  author = {{Carnegie Mellon Database Group}},
+  title = {{Database of Databases}},
+  journal = {Database of Databases},
+  year = {2022},
+  month = aug,
+  note = {[Online; accessed 1. Aug. 2022]},
+  url = {https://dbdb.io}
+}
+
+@misc{DBEngines,
+  author = {{solid IT GmbH}},
+  title = {{DB-Engines Ranking}},
+  journal = {DB-Engines},
+  year = {2022},
+  month = aug,
+  note = {[Online; accessed 1. Aug. 2022]},
+  url = {https://db-engines.com/en/ranking}
+}
+
 
 
