+++ paper.md
@@ -36,8 +36,9 @@ See the [homepage](https://github.com/Beuth-Erdelt/DBMS-Benchmarker) and the [do
 
 Performance benchmarking of database management systems (DBMS) is an active research area and has a broad audience. It is used *by DBMS developers to evaluate their work and to find out which algorithm works best in which situation. Benchmarks are used by (potential) customers to evaluate what system or hardware to buy or rent. Benchmarks are used by administrators to find bottlenecks and adjust configurations. Benchmarks are used by users to compare semantically equivalent queries and to find the best formulation alternative*, @10.1007/978-3-030-84924-5_6.
 Also in the academic field, approaches and their special implementations are examined in benchmarks.
-There are a variety of DBMS and a lot of products. For example 
-@DBEngines ranks 350 DBMS, @DBDBIO lists 850 DBMS.
+There are a variety of DBMS and a lot of products.
+For example @DBEngines ranks 350 DBMS (150 Relational), @DBDBIO lists 850 DBMS (280 Relational).
+We focus on Relational DBMS (RDBMS) in the following.
 The types thereof can be divided into for example row-wise, column-wise, in-memory, distributed and GPU-enhanced. 
 All of these products have unique characteristics, special use cases, advantages and disadvantages and their justification.
 In order to be able to verify and ensure the performance measurement, we want to be able to create and repeat benchmarking scenarios.
@@ -45,8 +46,8 @@ Repetition and thorough evaluation are crucial, in particular in the age of Clou
 
 Thus there is a widespread need for a tool to support the repetition and reproducibility of benchmarking situations, and that is capable of connecting to all these systems.
 
-There is also need for a tool that will help with the statistical, visual and interactive analysis of the results.
-We want to use Python as a common Data Science language, since
+When we collect a lot of data during benchmarking processes, we also need a tool that will help with the statistical, visual and interactive analysis of the results.
+The authors advocate using Python as a common Data Science language, since
 *it is a mature language programming, easy for the newbies, and can be used as a specific platform for data scientists, thanks to its large ecosystem of scientific libraries and its high and vibrant community*, @series/utcs/IgualS17.
 This helps to implement the tool into a pipeline, for example to make use of closed-loop benchmarking situations, @10114533389063338912, or to closely inspect parts of queries, @DBLPconfsigmodKerstenKZ18.
 It also allows to use common and sophisticated tools to inspect and evaluate the results.
