### 2022-08-01 Beuth-Erdelt/DBMS-Benchmarker#95 @perdelt
Hi @erik-whiting , thank you for your valuable hints!
I just revised the section *statement of need* and incorporated your suggestions for improvement.
I will push a new version.
