### 2022-08-01 a5d7ff4
V0.12.2 (#96)
* JOSS: Badge

* Docs: Runs with TimescaleDB

* Dashboard: Color by number of experiment run

* Dashboard: Client from Microservices

* Joss: Statement of need clearer and DOIs corrected

* Joss: Statement of need - DBs
