+++ paper.bib
@@ -156,7 +156,8 @@ Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and A
 Editor = {F. Loizides and B. Schmidt},
 Organization = {IOS Press},
 Pages = {87 - 90},
-Year = {2016}
+Year = {2016},
+doi = {10.3233/978-1-61499-649-1-87}
 }
 
 @Article{Hunter:2007,
@@ -213,3 +214,113 @@ Year = {2016}
   biburl    = {https://dblp.org/rec/books/sp/KounevLK20.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
+
+@book{series/utcs/IgualS17,
+  added-at = {2018-11-02T00:00:00.000+0100},
+  author = {Igual, Laura and Seguí, Santi},
+  biburl = {https://www.bibsonomy.org/bibtex/2df957256a471cf3626c5a280424a2b9a/dblp},
+  ee = {https://doi.org/10.1007/978-3-319-50017-1},
+  doi = {10.1007/978-3-319-50017-1},
+  interhash = {1f6d348e064db5ee3a118549b35cbb04},
+  intrahash = {df957256a471cf3626c5a280424a2b9a},
+  isbn = {978-3-319-50017-1},
+  keywords = {dblp},
+  pages = {1-215},
+  publisher = {Springer},
+  series = {Undergraduate Topics in Computer Science},
+  timestamp = {2018-11-03T12:45:44.000+0100},
+  title = {Introduction to Data Science - A Python Approach to Concepts, Techniques and Applications},
+  year = 2017
+}
+
+@article{Waskom2021,
+    doi = {10.21105/joss.03021},
+    url = {https://doi.org/10.21105/joss.03021},
+    year = {2021},
+    publisher = {The Open Journal},
+    volume = {6},
+    number = {60},
+    pages = {3021},
+    author = {Michael L. Waskom},
+    title = {seaborn: statistical data visualization},
+    journal = {Journal of Open Source Software}
+ }
+
+ @misc{TIOBE,
+  author = {TIOBE},
+  title = {{TIOBE Index - TIOBE}},
+  journal = {TIOBE},
+  year = {2022},
+  month = jun,
+  note = {[Online; accessed 31. Jul. 2022]},
+  url = {https://www.tiobe.com/tiobe-index}
+}
+
+@misc{PYPL,
+  author = {PYPL},
+  title = {{PYPL PopularitY of Programming Language index}},
+  year = {2022},
+  month = jul,
+  note = {[Online; accessed 31. Jul. 2022]},
+  url = {https://pypl.github.io/PYPL.html}
+}
+
+@inproceedings{10114533389063338912,
+author = {He, Sen and Manns, Glenna and Saunders, John and Wang, Wei and Pollock, Lori and Soffa, Mary Lou},
+title = {A Statistics-Based Performance Testing Methodology for Cloud Applications},
+year = {2019},
+isbn = {9781450355728},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3338906.3338912},
+booktitle = {Proceedings of the 2019 27th ACM Joint Meeting on European Software Engineering Conference and Symposium on the Foundations of Software Engineering},
+pages = {188--199},
+numpages = {12},
+keywords = {cloud computing, resource contention, performance testing, non- parametric statistics},
+location = {Tallinn, Estonia},
+series = {ESEC/FSE 2019},
+doi = {10.1145/3338906.3338912},
+}
+
+@inproceedings{DBLPconfsigmodKerstenKZ18,
+  author    = {Martin L. Kersten and
[…]
