### 2022-08-11 Beuth-Erdelt/DBMS-Benchmarker#97 @simon-lewis
JOSS Review - General checks
JOSS Review – General Checks by @simon-lewis

Summary: Completed. No issues to address or raised (subject to Contributing Guidelines on publication).

(1)	**Repository** – confirmed – sources available. No obvious issues.

(2)	**License** – visible and confirmed as AGPL-3.0 – OSI approved

(3)	**Contribution and authorship** – contributors confirmed. 336 commits for Patrick. 1 for Jascha. 100K+ lines contributed overall. No obvious issues except for missing contributing guidelines under Github Community Standards.

(4)	**Scholarly effort** – confirmed above. No citations evident but no DOI yet. Suggested benefits are real and the software is cite-able based on these General Checks (i.e. subject to review of Software Paper, Documentation and Functionality). Note close relationship with the referenced paper at [https://doi.org/10.1007/978-3-030-84924-5_6
](https://doi.org/10.1007/978-3-030-84924-5_6
)

Based on https://github.com/openjournals/joss-reviews/issues/4628
