### 2022-08-11 Beuth-Erdelt/DBMS-Benchmarker#98 @simon-lewis
JOSS Review - Software Paper
JOSS Review – Software Paper by @simon-lewis

**Summary**: Overall very good. Some suggested language changes under (4), offered for clarity and readability. 1 issue to address under (2) (points-of-difference / uniqueness).

**(1)	Statement of Need** – overall good but see section (4) below for suggested updates and corrections.

One issue that is not clear is based on the final paragraph (line 51 onwards in the reviewed version). This suggests the uniqueness of the software but it is not obvious how. For example, how does DBMS-Benchmarker differ from, say, HammerDB ?

I suspect its point-of-difference is in features like (a) randomisation of benchmarking queries to preclude caching, (b) its use of JDBC only, and/or (c) its long list of measurable DBMS.

If true that these – or other features - are indeed unique then in my opinion it is useful to state this in the final paragraph. That is, do not rely exclusively on citation links or an assertion.

**(2)	State of the field** – See (1) above.

**(3)	Quality of writing** – overall good. Line numbering for suggested changes refers to the PDF version published [here](https://github.com/openjournals/joss-reviews/issues/4628#issuecomment-1201246492) / https://github.com/openjournals/joss-reviews/issues/4628#issuecomment-1201246492

Line 19: start of paragraph is italicised. Not sure why. Suggest removing italics.
Line 25: suggest removing “and a lot of” – change sentence to be “There is a large variety of DBMS products” for readability.
Line 30: modify end of sentence to read “… and their own justification”.
Line 33: correct “it’s” to “its” (“it’s” is a contraction of “it is” which is not correct here)
Line 37: suggest removing “a lot of”
Line 39: middle of paragraph is italicised. Not sure why. Suggest removing italics.
Line 44: change “… allows to use common …” to “… allows the use of common …”
Line 55: correct spelling of “recieving” – should be “**receiving**”
Line 75: change “in parts” to “in part”
Line 77: remove “very” for readability
Line 120: remove comma (superfluous)
Line 151: recommend defining **timerSession** as not clear from Figure 1 or the text what the value represents (as distinct from **timerRun**, for example).
Line 154” change “… run and stops …” to “… runs and stops …”
Line 154: remove “been” as superfluous. End of sentence should read “… run has finished.”
Line 164: suggested changing last sentence beginning “This tool …” to “The tool automatically checks …”

[10.21105.joss.04628 (2).pdf](https://github.com/Beuth-Erdelt/DBMS-Benchmarker/files/9305669/10.21105.joss.04628.2.pdf)

**(4)	References** – look good to me. Difficult to be comprehensive in this space as there is (far too) much material out there. The references given seem relevant and useful, including those to Patrick’s related publications.

END.
