### 2022-08-13 Beuth-Erdelt/DBMS-Benchmarker#97 @perdelt
Hi @simon-lewis
many thanks for your checks!

As a reply: Jascha contributed quite a lot on the evaluation side. His commits have been squash-merged from another repository (so it counts as 1 here).
The paper https://doi.org/10.1007/978-3-030-84924-5_6 motivated the design decisions of the package and showed some results that have been gained using the package. Otherwise the content is disjoint. In particular terminology, internal logic and usage can only be found in the JOSS publication.
