### 2022-08-14 Beuth-Erdelt/DBMS-Benchmarker#98 @perdelt
Hi @simon-lewis

thank you for your useful remarks!

Some comments:
* Italicised means it is a quote. I added quotation marks to make this clearer.
* I kept “a lot of” in Line 37, since the need for evaluation tools increases with the amount of data collected.

I incorporated your other suggestions. I find them very helpful for improving readability. I also pushed a new version.
If you see more room for improvement, please let me know.
