### 2022-08-28 #4628 @erik-whiting
Howdy, I've completed all items in my initial review and for the most part, everything looks great. There are a few items in with the documentation that I think should be improved upon before publication; I've elaborated on these items on a [documentation-related issue](https://github.com/Beuth-Erdelt/DBMS-Benchmarker/issues/100) but I will generalize here:

1) I don't see any automated tests or instructions for manual tests. Can you please provide a way for people to verify functionality?

2) There is no contribution guide, can you please add one?

3) There is no information provided for people wanting to open issues or find support, can you please provide guidance for potential users that need help?

In the linked issue, there is further elaboration, just giving the synopsis here. Once these issues are addressed, it's a 👍 from me.

Oh and by the way, the documentation is great, very thorough!
