### 2022-08-28 Beuth-Erdelt/DBMS-Benchmarker#98 @simon-lewis
All good. Thanks Patrick. Happy with this one.
