### 2022-08-29 Beuth-Erdelt/DBMS-Benchmarker#101 @simon-lewis
JOSS Review - Documentation
JOSS Review - Documentation by @simon-lewis
Related: [JOSS](https://github.com/openjournals/joss-reviews/issues/4628)

Summary: Good README.md. Online documentation (Read the Docs.io @ [Docs](https://dbmsbenchmarker.readthedocs.io/en/latest/index.html)) adequate but with some limitations. No testing or contribution elements specified. The notes below refer to the Documentation @ readthedocs.io:

**(1)	A statement of need** – **Missing**. Introductory paragraph on README.md is good. Otherwise covered in the Software Paper (pedantically, there is no SoN in the main documentation @readthedocs but this is **not a major oversight**).

**(2) Installation instructions** – **Needs fixing**. Simple PIP INSTALL. Appeared to work on container, host and windows environments but subsequent execution (of dbmsbenchmarker) failed due to unfulfilled dependencies – these would not be an issue for an experienced user but worth being more detailed here. Dependencies like JAVA and the JDBC JAR file, whilst implied, should be explicit and clear in the documentation.

**(3)	Example usage** – Good documentation on sample usage with config examples, including simple and extended examples. Thorough.

**(4)	Functionality documentation** – good. No particular issues (see Functionality for related elements)

**(5) Automated tests** – **Missing**. none specified.

**(6) Community guidelines** – **Missing** information on Contributing.

END.
