### 2022-08-29 Beuth-Erdelt/DBMS-Benchmarker#102 @simon-lewis
JOSS Review – Functionality
JOSS Review – Functionality by @simon-lewis

[see JOSS](https://github.com/openjournals/joss-reviews/issues/4628)

**Summary**: good response from main script, noting there is a lag on startup which gets reflected in the results – it is possible this requires configuration of a connection value (see “**Connection Latency**” in the documentation - not tested).
Could not test _dbmsdashboard_ properly; possibly due to dependencies / permissions and/or local environment. Details below.

•	Hardware metrics not tested (Prometheus)
•	**dbmsdashboard** tested but failed, possibly due to local environment / un-met dependencies
•	Randomisation not tested
•	Connection Latency configuration not tested but could be significant as initial startup shows lag
•	Basic comparison testing completed between two MySQL schemas on single MySQL containerised instance. (Configs and DB scripts attached for info)
•	Tested within container with local MySQL instance
•	Tested in Linux VM with containerised MySQL instance
•	Tested in Windows PC to remote, containerised MySQL instance

Good results from scripts, with the following 3 example outputs included:

```
Q1: Count all rows in test - timerExecution
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+-------+
| DBMS [ms]   |     n |   mean |   stdev |   cv % |   qcod % |   iqr |   median |   min |   max |
+=============+=======+========+=========+========+==========+=======+==========+=======+=======+
| MySQL-1     | 10.00 |  10.23 |   18.78 | 183.57 |    11.33 |  0.85 |     3.74 |  3.15 | **63.35** |
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+-------+
| MySQL-2     | 10.00 |   3.61 |    1.07 |  29.76 |    20.57 |  1.45 |     3.03 |  2.68 |  5.57 |
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+-------+
Experiment 1661736663 has been finished

```

```
Q1: Count all rows in test - timerExecution
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+--------+
| DBMS [ms]   |     n |   mean |   stdev |   cv % |   qcod % |   iqr |   median |   min |    max |
+=============+=======+========+=========+========+==========+=======+==========+=======+========+
| MySQL-1     | 10.00 |  24.68 |   51.47 | 208.57 |    39.10 |  5.55 |     5.74 |  3.75 | **169.53** |
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+--------+
| MySQL-2     | 10.00 |   3.15 |    0.59 |  18.58 |     8.65 |  0.54 |     2.93 |  2.46 |   4.37 |
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+--------+
Experiment 1661736700 has been finished
```

```
Q1: Count all rows in test - timerExecution
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+-------+
| DBMS [ms]   |     n |   mean |   stdev |   cv % |   qcod % |   iqr |   median |   min |   max |
+=============+=======+========+=========+========+==========+=======+==========+=======+=======+
| MySQL-1     | 10.00 |  10.40 |   18.16 | 174.67 |    10.02 |  0.94 |     4.54 |  3.98 | **62.05 |**
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+-------+
| MySQL-2     | 10.00 |   3.82 |    0.75 |  19.74 |    15.02 |  1.10 |     3.96 |  2.77 |  4.92 |
+-------------+-------+--------+---------+--------+----------+-------+----------+-------+-------+
Experiment 1661736721 has been finished
```

Note the following – MAX values always come from initial query on first scheme (db01) so assumed to be a result of Connection Latency:

Run 1: `INFO:singleRun:numRun 1: execution [ms]: 63.35214083082974`
Run 2: `INFO:singleRun:numRun 1: execution [ms]: 169.5333649404347`
Run 3: `INFO:singleRun:numRun 1: execution [ms]: 62.05371394753456`

This skews the results so recommend documentation is updated to highlight this risk and - if it resolves the issue - to explain and recommend the use of Connection Latency values in the configuration.

`dbmsdashboard -r 1661736721`

Appeared to work but no results shown (note Permissions errors in output).
“(“PermissionError: [Errno 13] Permission denied: '/connections.config'”)”

END. Evaluation completed with the limitations / recommendations highlighted above.
