### 2022-08-29 Beuth-Erdelt/DBMS-Benchmarker#102 @simon-lewis
[Eval.zip](https://github.com/Beuth-Erdelt/DBMS-Benchmarker/files/9441511/Eval.zip)
