### 2022-08-29 #4628 @simon-lewis
Hi team,
My review is now completed. See issues raised under Docs and Functionality - mainly around startup / connection lag but nothing major. As @erik-whiting says, documentation is overall very good with great depth on example use-cases.
Thanks.
Simon
@simon-lewis
