### 2022-09-02 ea70bb4
default for all queries is being 'active'
