### 2022-09-02 9b137e7
Adjusted debug output to print again in subprocesses
