+++ dbmsbenchmarker/benchmarker.py
@@ -92,10 +92,11 @@ def singleRun(connectiondata, inputConfig, numRuns, connectionname, numQuery, pa
 	:return: returns object of class singleRunOutput
 	"""
 	#global activeConnections
-	import logging, multiprocessing
+	#import logging, multiprocessing
 	#logging.basicConfig(format='%(asctime)s %(message)s')
+	#logger = multiprocessing.log_to_stderr()
 	#logger = multiprocessing.get_logger()
-	logger = logging.getLogger('singleRun')
+	#logger = logging.getLogger('singleRun')
 	#logger.setLevel(logging.INFO)
 	# init list of results
 	results = []
@@ -110,7 +111,7 @@ def singleRun(connectiondata, inputConfig, numRuns, connectionname, numQuery, pa
 	#print(len(activeConnections))
 	if len(activeConnections) > numActiveConnection:
 		if BENCHMARKER_VERBOSE_PROCESS:
-			logger.info("Found active connection #"+str(numActiveConnection))
+			print("Found active connection #"+str(numActiveConnection))
 		connection = activeConnections[numActiveConnection]
 		durationConnect = 0.0
 	else:
@@ -118,7 +119,7 @@ def singleRun(connectiondata, inputConfig, numRuns, connectionname, numQuery, pa
 		query = tools.query(inputConfig[numRuns[0]].queryConfig)
 		if query.delay_connect > 0:
 			if BENCHMARKER_VERBOSE_PROCESS:
-				logger.info("Delay Connection by "+str(query.delay_connect)+" seconds")
+				print("Delay Connection by "+str(query.delay_connect)+" seconds")
 			time.sleep(query.delay_connect)
 		# connect to dbms
 		connection = tools.dbms(connectiondata)
@@ -127,8 +128,9 @@ def singleRun(connectiondata, inputConfig, numRuns, connectionname, numQuery, pa
 		end = default_timer()
 		durationConnect = 1000.0*(end - start)
 	if BENCHMARKER_VERBOSE_PROCESS:
-		logger.info(("singleRun batch size %i: " % len(numRuns)))
-		logger.info(("numRun %s: " % ("/".join([str(i+1) for i in numRuns])))+"connection [ms]: "+str(durationConnect))
+		print(("singleRun batch size %i: " % len(numRuns)))
+	if durationConnect > 0:
+		print(("numRun %s: " % ("/".join([str(i+1) for i in numRuns])))+"connection [ms]: "+str(durationConnect))
 	# perform runs for this connection
 	for numRun in numRuns:
 		workername = "numRun %i: " % (numRun+1)
@@ -137,7 +139,7 @@ def singleRun(connectiondata, inputConfig, numRuns, connectionname, numQuery, pa
 		query = tools.query(inputConfig[numRun].queryConfig)
 		if query.delay_run > 0:
 			if BENCHMARKER_VERBOSE_PROCESS:
-				logger.info(workername+"Delay Run by "+str(query.delay_run)+" seconds")
+				print(workername+"Delay Run by "+str(query.delay_run)+" seconds")
 			time.sleep(query.delay_run)
 		error = ""
 		try:
@@ -146,9 +148,9 @@ def singleRun(connectiondata, inputConfig, numRuns, connectionname, numQuery, pa
 				#logger.info(type(queryString))
 				if isinstance(queryString, list):
 					for queryPart in queryString:
-						logger.info(workername+queryPart)
+						print(workername+queryPart)
 				else:
-					logger.info(workername+queryString)
+					print(workername+queryString)
 			connection.openCursor()
 			#end = default_timer()
 			#durationConnect += 1000.0*(end - start)
[…]
