### 2022-09-03 Beuth-Erdelt/DBMS-Benchmarker#102 @perdelt
Hi @simon-lewis

Many thanks for your thorough and helpful tests!

You can start the dashboard by a simple `dbmsdashboard`. The result folder flag `-r` is only necessary if you also have used it during benchmarking. It points to the (parent) folder where the results should be put at. Otherwise the result folders are just put in the current path and `dbmsdashboard` looks there for them.

It is common to exclude the first run in benchmarking, because of the cold data / startup lag situation. I will add a comment to the documentation.

However I myself would like to see, what part of the effect is client-sided (the package) or server-sided (the DBMS). I will think about a test.

`timerExecution` for sure only starts measuring when a connection has been established.

You can (kind of) verify this by upgrading to the current version (I added some output infos and changed default behaviour), and use the query file attached.

```
{
  'name': 'Some simple queries',
  'connectionmanagement': {
        'timeout': 5, # in seconds
        'singleConnection': False, # not single connection per stream
        'runsPerConnection': 1     # reconnect for every execution
  },
  'queries':
  [
    {
      'title': "Static query",
      'query': "SELECT 1",
      'numRun': 10,
      'timer':
      {
        'connection':
        {
          'active': True,
          'delay': 0
        },
      }
    },
    {
      'title': "Count all rows in test",
      'query': "SELECT COUNT(*) FROM posts",
      'numRun': 10,
      'timer':
      {
        'connection':
        {
          'active': True,
          'delay': 0
        },
      }
    },
  ]
}
```

The `connectionmanagement` tells the benchmarker to reconnect for every execution of every query.
I will put more information about these setting in the documentation, too.

I then can see
```
numRun 1: connection [ms]: 1363.3746
numRun 1: execution [ms]: 156.53840000000008
numRun 2: connection [ms]: 305.8920999999999
numRun 2: execution [ms]: 31.420500000000295
numRun 3: connection [ms]: 306.9272999999999
numRun 3: execution [ms]: 31.359700000000323
numRun 4: connection [ms]: 289.99470000000025
numRun 4: execution [ms]: 30.865400000000598
numRun 5: connection [ms]: 290.58910000000003
numRun 5: execution [ms]: 30.43529999999972
numRun 6: connection [ms]: 292.0455999999998
numRun 6: execution [ms]: 30.581800000000214
numRun 7: connection [ms]: 299.7739999999993
numRun 7: execution [ms]: 33.33660000000016
numRun 8: connection [ms]: 314.39169999999984
numRun 8: execution [ms]: 31.637899999999775
numRun 9: connection [ms]: 304.4646000000002
numRun 9: execution [ms]: 33.20440000000069
numRun 10: connection [ms]: 318.0198000000001
numRun 10: execution [ms]: 33.112400000000264
```

so the first execution is slower and this has nothing to do with connecting.
