### 2022-09-03 Beuth-Erdelt/DBMS-Benchmarker#102 @perdelt
[queries.zip](https://github.com/Beuth-Erdelt/DBMS-Benchmarker/files/9482616/queries.zip)
