### 2022-09-03 88970c9
Docs: Contribution first draft
