+++ CONTRIBUTING.md
@@ -1,0 +1,42 @@
+# Contributing to DBMSBenchmarker
+
+You would like to contribute? Great!
+
+Some things that you can help on include:
+* **New Workloads**: The `example/` folder includes the TPC-H and TPC-DS (reading) queries in various dialects.
+* **Documentation**
+* **Testing**
+* **More DBMS**
+
+## Non-code contributions
+
+Even if you don’t feel ready or able to contribute code, you can still help out. There always things that can be improved on the documentation (even just proof reading, or telling us if a section isn’t clear enough).
+
+
+## Code contributions
+
+We welcome pull requests to fix bugs or add new features.
+
+### Licensing
+
+In your git commit and/or pull request, please explicitly state that you agree to your contributions being licensed under "GNU Affero General Public License v3".
+
+
+### Git Usage
+
+If you are planning to make a pull request, start by creating a new branch with a short but descriptive name (rather than using your master branch).
+
+
+### Coding Conventions
+
+DBMSBenchmarker tries to follow the coding conventions laid out in PEP8 and PEP257
+
+- http://www.python.org/dev/peps/pep-0008/
+- http://www.python.org/dev/peps/pep-0257/
+
+
+### Testing
+
+Any new feature or functionality will not be accepted without tests.
+Likewise for any bug fix, we encourage including an additional test.
+
