### 2022-09-04 103a026
Docs: Contribution second draft
