+++ docs/index.rst
@@ -18,7 +18,7 @@
     ./Evaluations.md
     ./Dashboard.md
     ./Inspection.md
-    ../CONTRIBUTING.md
+    ./CONTRIBUTING.md
     ...
 
 
