### 2022-09-05 decc840
Docs: Contribution third draft
