### 2022-09-05 5357692
Docs: singleConnection
