### 2022-09-05 e153069
Docs: Query and Connection als separate sections
