+++ docs/Options.md
@@ -113,83 +113,6 @@ Example: `-r /tmp/dbmsresults/`, and a subfolder, say `1234`, will be generated 
 Name of folder containing query and connection config files.
 If set, the names `connections.config` and `queries.config` are assumed automatically.
 
-### Connection File
-
-Contains infos about JDBC connections.
-
-Example for `CONNECTION_FILE`:
-```
-[
-  {
-    'name': "MySQL",
-    'version': "CE 8.0.13",
-    'info': "This uses engine innodb",
-    'active': True,
-    'alias': "DBMS A",
-    'docker': "MySQL",
-    'docker_alias': "DBMS A",
-    'dialect': "MySQL",
-    'timeload': 100,
-    'priceperhourdollar': 1.0,
-    'JDBC': {
-      'driver': "com.mysql.cj.jdbc.Driver",
-      'url': "jdbc:mysql://localhost:3306/database",
-      'auth': ["username", "password"],
-      'jar': "mysql-connector-java-8.0.13.jar"
-    },
-    'connectionmanagement': {
-      'timeout': 600,
-      'numProcesses': 4,
-      'runsPerConnection': 5
-    },
-    'hostsystem': {
-      'RAM': 61.0,
-      'CPU': 'Intel(R) Xeon(R) CPU E5-2686 v4 @ 2.30GHz\n',
-      'Cores': '8\n',
-      'host': '4.4.0-1075-aws\n',
-      'disk': '82G\n',
-      'CUDA': ' NVIDIA-SMI 410.79       Driver Version: 410.79       CUDA Version: 10.0',
-      'instance': 'p3.2xlarge'
-    },
-    'monitoring': {
-      'shift': 0,
-      'extend': 20,
-      'prometheus_url': 'http://127.0.0.1:9090/api/v1/',
-      'metrics': {
-        'total_cpu_memory': {
-          'query': 'container_memory_working_set_bytes{job="monitor-node"}/1024/1024',
-          'title': 'CPU Memory [MiB]'
-        }
-      }
-    }
-},
-]
-```
-
-* `name`: References the connection
-* `version` and `info`: Just info texts for implementation in reports
-* `active`: Use this connection in benchmarking and reporting (optional, default True)
-* `alias`: Alias for anonymized reports (optional, default is a random name)
-* `docker`: Name of the docker image. This helps aggregating connections using the same docker image.
-* `docker_alias`: Anonymized name of the docker image. This helps aggregating connections using the same docker image in anonymized reports.
-* `alias`: Alias for anonymized reports (optional default is a random name)
-* `dialect`: Key for (optional) alternative SQL statements in the query file
-* `driver`, `url`, `auth`, `jar`: JDBC data
-* Additional information useful for reporting and also used for computations
-  * `timeload`: Time for ingest (in milliseconds), because not part of the benchmark
-  * `priceperhourdollar`: Used to compute total cost based on total time (optional)
-* `connectionmanagement`: Parameter for connection management. This overwrites general settings made in the [query config](#extended-query-file) and can be overwritten by query-wise settings made there.
-  * `timeout`: Maximum lifespan of a connection. Default is None, i.e. no limit.
-  * `numProcesses`: Number of parallel client processes. Default is 1.
-  * `runsPerConnection`: Number of runs performed before connection is closed. Default is None, i.e. no limit.
-  * `singleConnection`: This indicates if the connection should be used for the complete stream of queries. Default is True. Switch this off, if you want to have reconnects during the stream, for example to inspect the effect of reconnection of execution times.
-* `hostsystem`: Describing information for report in particular about the host system.
-  This can be written automatically by https://github.com/Beuth-Erdelt/Benchmark-Experiment-Host-Manager
[…]
