### 2022-09-08 Beuth-Erdelt/DBMS-Benchmarker#101 @perdelt
Hi @simon-lewis, thank you for your review!

I fixed the following
* (2) included requirements concerning JAVA
* (6) link to bug tracker and contribution guidelines (good idea btw!)

I'm unsure how to provide automated tests. Without a DBMS, the tool makes no sense. Do you have any idea? What would be appropriate? A check that certain results would be expected if installed correctly? Something like exit code = 0?
