### 2022-09-08 Beuth-Erdelt/DBMS-Benchmarker#100 @erik-whiting
> A check that certain results would be expected if installed correctly? Something like exit code = 0?

I think this would be perfect for users. Regarding contributors:

Imagine I opened a pull-request against DBMS-Benchmarker, what steps would you go through to make sure my PR didn't break the whole system? When I learned this term, it was called a "smoke test", I don't know if that's still the common terminology. But basically, is there a series of steps you go through to make sure the software hasn't broken since the last time you updated it? If so, document that (maybe "a check that cerrtain results would be expected if installed correctly" is good enough).

Does that help?
