### 2022-09-26 Beuth-Erdelt/DBMS-Benchmarker#102 @simon-lewis
Thanks Patrick. All good. Sorry for delayed response but I have been away for 3 week.

I'll close this as responded.

Just to note that the dashboard problems were probably due to my environment - **it is not an issue for this review** - but I did run it as you and the documentation recommends. It is just that I could not get it to function in any of my environments.
