### 2022-09-26 Beuth-Erdelt/DBMS-Benchmarker#101 @simon-lewis
Great.

(1) Tests - yes, I can see this is a challenge. To be clear, what JOSS (and indeed, Github) are seeking here are standard tests which any contributor can employ to verify their code contributions. That is, how does a programmer prove they have not 'broken' DBMS-Benchmarker by changing / adding new source code ? Difficult to offer detailed advice here but my suggestion is to consider a test script / procedure that could be followed by any developer to prove the code operates completely and reliably BUT without being dependant on a specific DBMS.

- Are all timers producing results ?
- Are those results reasonable and consistent ?
- Are those results accurate ? (I can't think of a sensible way to produce parallel / independent timings to compare with the results which would prove that the timers remain accurate, though. You may know of a way.)

Something to think about longer term. Perhaps take a look at other OS projects to see how they evolve generic test packs for their project contributors to use.

(2) Contributions statement - perfect. This is what JOSS is seeking. Thanks.
