### 2022-10-04 Beuth-Erdelt/DBMS-Benchmarker#101 @perdelt
Hi @simon-lewis
Thank you for the valuable hints!
I will consider adding a de-facto standard test setup beyond 'use TPC-H'. 'before-and-after' testing sounds very reasonable.

As this issue is closed, do you expect something more from my side right now before review can continue?
