### 2022-10-16 dce7e40
Removed deprecated reporters
