+++ dbmsbenchmarker/reporter.py
@@ -411,1926 +411,3 @@ class metricer(reporter):
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-class DEPRECATED_ploter(reporter):
-	"""
-	Class for generating reports.
-	Generates a plot of the benchmarks as a time series and saves it to disk.
-	"""
-	def __init__(self, benchmarker):
-		reporter.__init__(self, benchmarker)
-		self.e = None
-	def save(self, dataframe, query, title, subtitle, filename, timer):
-		"""
-		Saves report of a given query as plot image per timer.
-		Anonymizes dbms if activated.
-
-		:param dataframe: Report data given as a pandas DataFrame (rows=dbms, cols=benchmarks)
-		:param query: Query object
-		:param title: Title of the report
-		:param subtitle: Subtitle of the report
-		:param filename: Name of the file the report will be saved to
-		:return: returns nothing
-		"""
-		# test if any rows left
-		if len(dataframe.index) < 1:
-			return False
-		# transpose
-		df_transposed = dataframe.transpose()
-		# add index name
-		df_transposed.index.name = subtitle
-		# plot
-		#print(tools.dbms.dbmscolors)
-		plotdata = df_transposed.plot(title=title, figsize=(12,8), color=[tools.dbms.dbmscolors.get(x, '#333333') for x in df_transposed.columns])
-		# scale y-axis: 0 to max plus 10%
-		plotdata.set_ylim(0, df_transposed.max().max()*1.10,0)
-		if timer.perRun:
-			if query.warmup > 0:
-				plt.axvline(x=query.numRunBegin+1, linestyle="--", color="black")
-			if query.cooldown > 0:
-				plt.axvline(x=query.numRunEnd, linestyle="--", color="black")
-		plt.legend(title="DBMS")
-		plt.savefig(filename, bbox_inches='tight')
-		plt.close('all')
-	def generate(self, numQuery, timer):
-		"""
-		Generates a plot of the benchmarks as a time series and saves it to disk.
-		Anonymizes dbms if activated.
-
-		:param numQuery: Number of query to generate report of
-		:param timer: Timer containing benchmark results
-		:return: returns nothing
-		"""
-		for t in timer:
-			# are there benchmarks for this query?
-			if not t.checkForSuccessfulBenchmarks(numQuery):
-				continue
-			query = tools.query(self.benchmarker.queries[numQuery-1])
-			# is timer active for this query?
-			if not query.timer[t.name]['active']:
-				continue
-			# is query active?
-			if not query.active:
-				continue
-			# benchmark times as a dataframe
-			#df = self.benchmarker.benchmarksToDataFrame(numQuery, t)
-			#print(df)
-			if self.e is None:
-				self.e = evaluator.evaluator(self.benchmarker)
-			df = evaluator.dfMeasuresQ(numQuery, t.name)
-			#df.index = df.index.map(tools.dbms.anonymizer)
-			#print(df)
-			logging.debug("Plot Q"+str(numQuery)+" for timer "+t.name)
-			if t.perRun:
-				#subtitle = "Warmup = "+str(query.warmup)+", test runs = "+str(query.numRun-query.warmup-query.cooldown)+", cooldown = "+str(query.cooldown)
-				subtitle = "Number of run"
-			else:
-				subtitle = "Number of session"
-			# save as plot
-			self.save(
-				dataframe = df,
-				query = query,
-				#title = "Q"+str(numQuery)+": Time "+t.name+" [ms]",
[…]
