### 2022-10-16 e4c657a
Removed deprecated reporters (#105)
* Docs: Contribution tests

* Removed deprecated reporters
