+++ demo-interactive-benchmarking.py
@@ -17,10 +17,6 @@ dbms = benchmarker.benchmarker(result_path=resultfolder, numProcesses=8)
 dbms.getConfig(configfolder)
 dbms.reporter.append(benchmarker.reporter.pickler(dbms))
 dbms.reporter.append(benchmarker.reporter.dataframer(dbms))
-dbms.reporter.append(benchmarker.reporter.barer(dbms))
-dbms.reporter.append(benchmarker.reporter.ploter(dbms))
-dbms.reporter.append(benchmarker.reporter.boxploter(dbms))
-dbms.reporter.append(benchmarker.reporter.latexer(dbms, "pagePerQuery"))
 dbms.runBenchmarks()
 #dbms.continueBenchmarks(overwrite = False)
 #dbms.readBenchmarks()
