### 2022-10-16 20b67d0
Fixed dependencies to more current versions
