+++ requirements.txt
@@ -14,20 +14,17 @@ chardet>=3.0.4
 click>=6.7
 colour>=0.1.5
 cycler>=0.10.0
-dash==2.3.0
-dash-auth==1.4.1
-dash-core-components==2.0.0
-dash-daq==0.5.0
-dash-html-components==2.0.0
-dash-renderer==1.4.1
-dash-table==5.0.0
-Flask>=1.1.2
-Flask-Caching>=1.9.0
-Flask-Compress>=1.5.0
+dash==2.6.2
+dash-auth
+dash-core-components
+dash-daq
+dash-html-components
+dash-renderer
+dash-table
+Flask==2.1.0
 future>=0.18.2
 idna>=2.9
 itsdangerous>=1.1.0
-Jinja2==2.11.3
 kiwisolver>=1.2.0
 MarkupSafe>=1.1.1
 plotly>=4.8.1
@@ -38,7 +35,7 @@ retrying>=1.3.3
 six>=1.14.0
 urllib3==1.26.5
 uuid>=1.30
-Werkzeug>=1.0.1
+Werkzeug==2.1.0
 m2r2
 myst_parser
 markupsafe==2.0.1
