+++ MANIFEST.in
@@ -1,3 +1,2 @@
 include LICENSE
-include dbmsbenchmarker/latex/pagePerQuery/*
 recursive-include dbmsbenchmarker/scripts/assets *
