### 2022-10-16 1710f8d
Fixed dependencies to more current versions
