### 2022-10-16 e330502
Fixed dependencies to more current versions (#106)
* Fixed dependencies to more current versions
