### 2022-10-17 #4628 @perdelt
Hi @gkthiruvathukal please find the information here

> Hi @perdelt. Thank you for your patience. Please do the following:
>
> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
  **v0.13.0**
> * [x]  Archive the reviewed software in Zenodo
> * [x]  Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
> * [x]  List the Zenodo DOI of the archived version here.
  **10.5281/zenodo.7213676**
