### 2022-11-01 #4628 @danielskatz
In this case, the errors seem to be a figure label that is reused (in the .md file), and two reference labels that are reused (in the .bib file) - the author will need to address these
