+++ paper.bib
@@ -83,23 +83,6 @@ doi = {10.1007/978-3-030-12079-5_4}
  doi = {10.1145/3209950.3209955}
 }
 
-@inproceedings{10114533389063338912,
-author = {He, Sen and Manns, Glenna and Saunders, John and Wang, Wei and Pollock, Lori and Soffa, Mary Lou},
-title = {A Statistics-Based Performance Testing Methodology for Cloud Applications},
-year = {2019},
-isbn = {9781450355728},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-url = {https://doi.org/10.1145/3338906.3338912},
-booktitle = {Proceedings of the 2019 27th ACM Joint Meeting on European Software Engineering Conference and Symposium on the Foundations of Software Engineering},
-pages = {188--199},
-numpages = {12},
-keywords = {cloud computing, resource contention, performance testing, non- parametric statistics},
-location = {Tallinn, Estonia},
-series = {ESEC/FSE 2019},
-doi = {10.1145/3338906.3338912}
-}
-
 @inproceedings{DBLPconfsigmodKerstenKZ18,
   author    = {Martin L. Kersten and
                Panagiotis Koutsourakis and
@@ -280,25 +263,6 @@ keywords = {cloud computing, resource contention, performance testing, non- para
 location = {Tallinn, Estonia},
 series = {ESEC/FSE 2019},
 doi = {10.1145/3338906.3338912},
-}
-
-@inproceedings{DBLPconfsigmodKerstenKZ18,
-  author    = {Martin L. Kersten and
-               Panagiotis Koutsourakis and
-               Ying Zhang},
-  editor    = {Alexander B{\"{o}}hm and
-               Tilmann Rabl},
-  title     = {Finding the Pitfalls in Query Performance},
-  booktitle = {Proceedings of the 7th International Workshop on Testing Database
-               Systems, DBTest@SIGMOD 2018, Houston, TX, USA, June 15, 2018},
-  pages     = {3:1--3:6},
-  publisher = {{ACM}},
-  year      = {2018},
-  url       = {https://doi.org/10.1145/3209950.3209951},
-  doi       = {10.1145/3209950.3209951},
-  timestamp = {Mon, 12 Aug 2019 13:49:51 +0200},
-  biburl    = {https://dblp.org/rec/conf/sigmod/KerstenKZ18.bib},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
 
