### 2022-11-01 #4628 @perdelt
> In this case, the errors seem to be a figure label that is reused (in the .md file), and two reference labels that are reused (in the .bib file) - the author will need to address these

Hi @gkthiruvathukal , Hi @danielskatz thanks for pointing out these issues!
I have corrected both files in the master branch (do I need a new tag?).
