### 2022-11-01 #4628 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
