### 2022-11-01 #4628 @danielskatz
> I don't think a new tag is needed when the issue is limited to the paper itself. Let's see.

Correct. Neither a new archive or tag, though both are allowed if desired. Since once the paper is published, the pdf is archived and the source no longer is important.
