### 2022-11-01 #4628 @danielskatz
@perdelt - as the track editor here, I'm doing the final checks before publishing.  I've found a number of issues with the final paper, as suggested in https://github.com/Beuth-Erdelt/DBMS-Benchmarker/pull/109.  Please merge this, or let me know what you disagree with, then we can proceed
