### 2022-11-02 ff6615e
changes in paper and bib (#109)
* changes in paper

* minor changes in bib

* undo Data Science edit
