+++ paper.md
@@ -22,48 +22,47 @@ bibliography: paper.bib
 # Summary
 
 DBMS-Benchmarker is a Python-based application-level blackbox benchmark tool for Database Management Systems (DBMS).
-It aims at reproducible measuring and easy evaluation of the performance the user receives even in complex benchmark situations.
+It is intended for reproducible measurement and easy evaluation of the performance the user receives, even in complex benchmark situations.
 It connects to a given list of DBMS (via JDBC) and runs a given list of (SQL) benchmark queries.
 Queries can be parametrized and randomized.
 Results and evaluations are available via a Python interface and can be inspected with standard Python tools like pandas DataFrames.
 An interactive visual dashboard assists in multi-dimensional analysis of the results.
 
-This module has been tested with Clickhouse, Exasol, Citus Data (Hyperscale), IBM DB2, MariaDB, MariaDB Columnstore, MemSQL (SingleStore), MonetDB, MySQL, OmniSci (HEAVY.AI) , Oracle DB, PostgreSQL, SQL Server, SAP HANA, TimescaleDB and Vertica.
+This module has been tested with Clickhouse, Exasol, Citus Data (Hyperscale), IBM DB2, MariaDB, MariaDB Columnstore, MemSQL (SingleStore), MonetDB, MySQL, OmniSci (HEAVY.AI), Oracle DB, PostgreSQL, SQL Server, SAP HANA, TimescaleDB, and Vertica.
 
-See the [homepage](https://github.com/Beuth-Erdelt/DBMS-Benchmarker) and the [documentation](https://dbmsbenchmarker.readthedocs.io/en/latest/Docs.html).
+See the [homepage](https://github.com/Beuth-Erdelt/DBMS-Benchmarker) and the [documentation](https://dbmsbenchmarker.readthedocs.io/en/latest/Docs.html) for more details.
 
 # Statement of Need
 
 Performance benchmarking of database management systems (DBMS) is an active research area and has a broad audience. It is used "*by DBMS developers to evaluate their work and to find out which algorithm works best in which situation. Benchmarks are used by (potential) customers to evaluate what system or hardware to buy or rent. Benchmarks are used by administrators to find bottlenecks and adjust configurations. Benchmarks are used by users to compare semantically equivalent queries and to find the best formulation alternative*", @10.1007/978-3-030-84924-5_6.
-Also in the academic field, approaches and their special implementations are examined in benchmarks.
+Approaches and their special implementations are also examined in benchmarks in academia.
 There is a large variety of DBMS types and products.
-For example @DBEngines ranks 350 DBMS (150 Relational), @DBDBIO lists 850 DBMS (280 Relational).
+For example, @DBEngines ranks 350 DBMS (150 Relational) and @DBDBIO lists 850 DBMS (280 Relational).
 We focus on Relational DBMS (RDBMS) in the following.
-The types thereof can be divided into for example row-wise, column-wise, in-memory, distributed and GPU-enhanced. 
+Their types can be divided into, for example, row-wise, column-wise, in-memory, distributed, and GPU-enhanced. 
 All of these products have unique characteristics, special use cases, advantages and disadvantages, and their own justification.
-In order to be able to verify and ensure the performance measurement, we want to be able to create and repeat benchmarking scenarios.
-Repetition and thorough evaluation are crucial, in particular in the age of Cloud-based systems with its diversity of hardware configurations, @Raasveldt2018FBC32099503209955, @DBLPconfsigmodKerstenKZ18, @KounevLK20.
+In order to be able to verify and ensure performance measurements, we want to be able to create and repeat benchmarking scenarios.
[…]
