+++ paper.bib
@@ -90,8 +90,8 @@ doi = {10.1007/978-3-030-12079-5_4}
   editor    = {Alexander B{\"{o}}hm and
                Tilmann Rabl},
   title     = {Finding the Pitfalls in Query Performance},
-  booktitle = {Proceedings of the 7th International Workshop on Testing Database
-               Systems, DBTest@SIGMOD 2018, Houston, TX, USA, June 15, 2018},
+  booktitle = {Proceedings of the 7th {I}nternational {W}orkshop on {T}esting {D}atabase
+               {S}ystems, DBTest@SIGMOD 2018,},
   pages     = {3:1--3:6},
   publisher = {{ACM}},
   year      = {2018},
@@ -212,7 +212,7 @@ doi = {10.3233/978-1-61499-649-1-87}
   publisher = {Springer},
   series = {Undergraduate Topics in Computer Science},
   timestamp = {2018-11-03T12:45:44.000+0100},
-  title = {Introduction to Data Science - A Python Approach to Concepts, Techniques and Applications},
+  title = {Introduction to Data Science - A {P}ython Approach to Concepts, Techniques and Applications},
   year = 2017
 }
 
