### 2022-11-02 #4628 @danielskatz
@perdelt - I think this looks ok, but since there were a large number of changes, you can also verify that the proof looks ok to you, before we proceed to publishing?
